采用静态方法处理，留下没有技术的泪水

+++ boot-core/src/main/java/top/huzhurong/test/bootcore/BeanMethodRegister.java
@@ -3,23 +3,24 @@ package top.huzhurong.test.bootcore;
 import top.huzhurong.test.bootcore.bean.BeanInfo;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * @author chenshun00@gmail.com
  * @since 2019/3/26
  */
 public class BeanMethodRegister {
-    private final static AtomicInteger count = new AtomicInteger(0);
+    private final static AtomicLong count = new AtomicLong(0);
 
     private static BeanInfo[] beanInfos = new BeanInfo[1024];
 
-    public synchronized static int hookKey(String className, String method) {
-        int key = count.incrementAndGet();
-        beanInfos[key] = new BeanInfo(className, method);
+    public synchronized static long hookKey(String className, String method) {
+        long key = count.incrementAndGet();
+        beanInfos[(int) key] = new BeanInfo(className, method);
         return key;
     }
 
-    public static BeanInfo get(int key) {
-        return beanInfos[key];
+    public static BeanInfo get(long key) {
+        return beanInfos[(int) key];
     }
 }
